Support multiple IPv6 prefixes on internal router ports
(Patch set #3 for the multiple-ipv6-prefixes blueprint)

Provides support for adding multiple IPv6 subnets to an internal router port.
The limitation of one IPv4 subnet per internal router port remains,
though a port may contain one IPv4 subnet with any number of IPv6
subnets.

This changes the behavior of both the router-interface-add and
router-interface-delete APIs. When router-interface-add is called with an
IPv6 subnet, the subnet will be added to an existing internal port on the
router with the same network ID if the existing port already has one or
more IPv6 subnets. Otherwise, a new port will be created on the router
for that subnet. When calling the router-interface-add with a port (one
that has already been created using the port-create command), that port
will be added to the router if it meets the following conditions:

        1. The port has no more than one IPv4 subnet.
        2. If the port has any IPv6 subnets, it must not have the same
           network ID as an existing port on the router if the existing
           port has any IPv6 subnets.

If the router-interface-delete command is called with a subnet, that
subnet will be removed from the router port to which it belongs. If the
subnet is the last subnet on a port, the port itself will be deleted
from the router. If the router-interface-delete command is called with a
port, that port will be deleted from the router.

This change also allows the RADVD configuration to support advertising multiple
prefixes on a single router interface.

This patchset does not include the following work items for the
'multiple-ipv6-prefixes' blueprint:
- Port-create API handling changes (proposed in earlier patchset)
- Port-update API handling changes (proposed in earlier patchset)
- Dual-stack external gateway ports (proposed in earlier patchset)
- Subnet create changes (TBD soon)

Change-Id: I7d4e8194815e626f1cfa267f77a3f2475fdfa3d1
Partially-implements: blueprint multiple-ipv6-prefixes