No allocation needed for specific IPv6 SLAAC addr assignment
(Patch set #7 for the multiple-ipv6-prefixes blueprint)

This patch set prevents calls to _allocate_specific_ip for
requests of a specific IPv6 SLAAC address assignment on an
internal router port. In this case, all that is required
is a uniqueness check on the requested address, as is
already being done in _test_fixed_ips_for_port.

Change-Id: I2533ee82980bb602faa663b875787ca50b268b34
Partially-implements: blueprint multiple-ipv6-prefixes